aclic: Phase 2 — widen mithreshold/sithreshold to 9-bit, extend mistatus layout
Per spec §Smnip "New Interrupt enable threshold":
- mithreshold.iprio / sithreshold.iprio is IPRIOLEN+1 bits (9-bit max),
  so values >= 2^IPRIOLEN ("all enabled") are now representable.
- Aclic::setMithreshold/setSithreshold now take uint16_t; private fields
  widened from uint8_t to uint16_t; CSR write masks widened to 0x1FF.
- Aclic::topInterrupt collapses threshold >= 2^IPRIOLEN to 0
  (no suppression), matching spec semantics.

Per spec §Smnip "New Interrupt Status", mistatus/sistatus layout updated:
- [16:8]  pithreshprio (9-bit, was [7:0] 8-bit before)
- [7]     psppush, [6] sppush (when Smcsps)
- [29:28] mpp / [28] spp        — alias of {mstatus.mpp, sstatus.spp}
- [27]    mpie / spie           — alias of xstatus.xpie
- [26]    mpelp / spelp         — alias of xstatus.xpelp (mstatush[9] on RV32)
- [20:19] VS, [18:17] FS        — alias of xstatus
- [0]     mie / sie             — alias of xstatus.xie

CSR read/write handlers in CsRegs.cpp assemble/disassemble the aliased
fields against (m|s)status; the owned pithreshprio+sppush bits live in
the CSR's backing store.  Trap entry (Hart.cpp:3676/3721) and xret
restore (Hart.cpp:12161/12247/12349) updated to use the new field
positions and 9-bit width.

+++ Aclic.hpp
@@ -22,11 +22,12 @@ public:
     bool hasSupervisorDomain() const { return hasSupervisorDomain_; }
     unsigned ipriolen() const { return ipriolen_; }
 
-    // Threshold CSRs (mithreshold/sithreshold). Values are masked to ipriolen bits.
-    void setMithreshold(uint8_t val);
-    void setSithreshold(uint8_t val);
-    uint8_t getMithreshold() const { return mithreshold_; }
-    uint8_t getSithreshold() const { return sithreshold_; }
+    // Threshold CSRs (mithreshold/sithreshold). Per spec §Smnip the field is
+    // IPRIOLEN+1 bits wide (9-bit max).  Values are masked to (ipriolen+1) bits.
+    void setMithreshold(uint16_t val);
+    void setSithreshold(uint16_t val);
+    uint16_t getMithreshold() const { return mithreshold_; }
+    uint16_t getSithreshold() const { return sithreshold_; }
 
     // Preemption mask (mipreemptcfg.preemptmsk). Controls which bits of the
     // priority participate in threshold comparison (NIPPRIO_MASK = ~(2^n - 1)).
@@ -92,9 +93,10 @@ private:
     unsigned numSources_;
     bool hasSupervisorDomain_;
     unsigned ipriolen_;
-    uint8_t mithreshold_ = 0;
-    uint8_t sithreshold_ = 0;
-    uint8_t mPreemptmsk_ = 0;  // mipreemptcfg.preemptmsk (0 = no masking)
+    uint16_t mithreshold_ = 0;     // 9-bit (IPRIOLEN+1) per Smnip spec
+    uint16_t sithreshold_ = 0;     // 9-bit
+    uint8_t mPreemptmsk_ = 0;      // legacy mipreemptcfg.preemptmsk; replaced by
+                                   // miconfig.{mnipbits,snipbits} in Phase 4
     DeliveryCallback deliveryCb_;
 
     void updateDelivery(bool isMachine);
